Refactor to share
