+++ tools/procgen.php
@@ -8,23 +8,24 @@ $procgen = new ProcGenerator(__DIR__ . '/../');
 
 ob_start();
 
-echo $procgen->preamble();
+echo $procgen->preamble('Processors');
 
 foreach ($procgen->getProcessorDetails() as $ii => $details) {
-    echo PHP_EOL . '## ' . $details['name'] . PHP_EOL;
-    echo 'Default: ' . ($details['default'] ? 'yes' : 'no') . ($details['default'] ? "(# $ii)" : '') . PHP_EOL;
+    $off = $ii + 1;
+    echo $procgen->formatClassHeader($details['name'], 'Processors');
     if ($details['properties']) {
-        echo '### Properties' . PHP_EOL;
+        $configPrefix = lcfirst($details['name']).'.';
+        echo '### Config settings' . PHP_EOL;
         foreach ($details['properties'] as $name => $type) {
             $var = ' : <span style="font-family: monospace;">' . $type . '</span>';
             echo '<dl>' . PHP_EOL;
-            echo '  <dt><strong>' . $name . '</strong>' . $var . '</dt>' . PHP_EOL;
+            echo '  <dt><strong>' . $configPrefix.$name . '</strong>' . $var . '</dt>' . PHP_EOL;
             echo '  <dd>';
             echo '<p>' . nl2br(ProcGenerator::NO_DETAILS_AVAILABLE) . '</p>' . PHP_EOL;
             echo '  </dd>' . PHP_EOL;
             echo '</dl>' . PHP_EOL;
         }
-
+        echo PHP_EOL;
     }
 }
