+++ tools/refgen.php
@@ -10,8 +10,10 @@ foreach ($refgen->types() as $type) {
     ob_start();
 
     echo $refgen->preamble($type);
+    echo 'In addition to this page, there are also a number of [examples](https://github.com/zircote/swagger-php/tree/master/Examples#readme) which might help you out.' . PHP_EOL . PHP_EOL;
+
     foreach ($refgen->classesForType($type) as $name => $details) {
-        echo $refgen->formatHeader($name, $type);
+        echo $refgen->formatClassHeader($name, $type);
         $method = "format{$type}Details";
         echo $refgen->$method($name, $details['fqdn'], $details['filename']);
     }
